+++ lib/shared/functions/quick_scan.dart
@@ -4,11 +4,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:http/http.dart';
 import 'package:receipt_wrangler_mobile/api/api.dart' as api;
+import 'package:receipt_wrangler_mobile/receipts/widgets/quick_scan.dart';
 import 'package:receipt_wrangler_mobile/shared/widgets/bottom_submit_button.dart';
 import 'package:receipt_wrangler_mobile/utils/scan.dart';
 import 'package:receipt_wrangler_mobile/utils/snackbar.dart';
 
-import '../../receipts/widgets/quick_scan_form.dart';
 import '../../utils/bottom_sheet.dart';
 
 Widget _getUploadIcon(
@@ -67,7 +67,7 @@ showQuickScanBottomSheet(context) {
 
   showFullscreenBottomSheet(
       context,
-      QuickScanForm(
+      QuickScan(
         formKey: formKey,
         imageStream: streamController.stream,
       ),
